requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
httpie==2.4.0
idna==2.10
Naked==0.1.31
pkg-resources==0.0.0
Pygments==2.8.1
PyJWT==2.0.1
PySocks==1.7.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-toolbelt==0.9.1
sqlparse==0.4.1
status==0.2.5
urllib3==1.26.4
